better fix

+++ pages/api/fix.ts
@@ -4,31 +4,48 @@ import { deleteEmailFromOutbox, getOutbox } from 'src/utils/api/email-engine';
 import { supabase } from 'src/utils/supabase-client';
 
 const _fixStuckInSequenceWithNoEmail: NextApiHandler = async (req, res) => {
-    const _company_id = '12513ff1-c7d9-417c-9e38-7e9d8de459bd';
+    const _company_ids = [
+        '504528ad-8f57-45e2-b82c-59ceb4bc9c54',
+        '9ffda880-addd-4d28-b247-92b1884a3cd9',
+        '625aa3ea-a54e-436a-a4c6-89ee09b148bb',
+        'd9d76bb4-9fe5-4ddb-bf1b-83eea7505281',
+        '2d01d750-8668-4990-812f-38b36c35e8c3',
+        'ea8ad9fb-7c52-46c3-bd95-4709b1dc70a5',
+    ];
     console.log('fixing');
     const outbox = await getOutbox();
 
     const { data: allSequenceInfluencers } = await supabase
         .from('sequence_influencers')
-        .select('*')
+        .select('id, name, company_id')
         .eq('funnel_status', 'In Sequence')
-        .eq('company_id', _company_id);
+        .in('company_id', _company_ids);
     console.log('allSequenceInfluencers', allSequenceInfluencers?.length);
     if (!allSequenceInfluencers) return res.status(200).json({ message: 'ok' });
     const updated: any = [];
     for (const influencer of allSequenceInfluencers) {
         const {
-            data: email,
+            data: emails,
             count,
             error: _error,
         } = await supabase
             .from('sequence_emails')
-            .select('id, email_message_id', { count: 'exact' })
+            .select('email_message_id, email_delivery_status, email_tracking_status', { count: 'exact' })
             .eq('sequence_influencer_id', influencer.id);
-        console.log('emails', count, '   error: ', _error);
+        // console.log('emails', count, '   error: ', _error);
 
-        if (email && count !== null && count < 4) {
-            const messageIds = email.map((e) => e.email_message_id);
+        if (emails && count !== null && count < 4) {
+            const hasAlreadyDelivered = emails.some(
+                (e) =>
+                    e.email_delivery_status === 'Delivered' ||
+                    e.email_tracking_status === 'Link Clicked' ||
+                    e.email_tracking_status === 'Opened',
+            );
+            if (hasAlreadyDelivered) {
+                console.log('hasAlreadyDelivered', influencer.id, influencer.company_id);
+                continue;
+            }
+            const messageIds = emails.map((e) => e.email_message_id);
             console.log('updating, ', influencer.name, influencer.company_id);
             await supabase.from('sequence_influencers').update({ funnel_status: 'To Contact' }).eq('id', influencer.id);
             updated.push(influencer.name + ' ' + influencer.id);
@@ -38,14 +55,16 @@ const _fixStuckInSequenceWithNoEmail: NextApiHandler = async (req, res) => {
                 console.log(`canceled ${email} ${canceledConfirm.deleted}`);
             }
 
-            console.log('deleting', email);
-            const { data: deleteConfirm } = await supabase
-                .from('sequence_emails')
-                .delete()
-                .eq('sequence_influencer_id', influencer.id)
-                .select('id');
+            // console.log('deleting', emails);
+            if (emails.length > 0) {
+                const { data: deleteConfirm } = await supabase
+                    .from('sequence_emails')
+                    .delete()
+                    .eq('sequence_influencer_id', influencer.id)
+                    .select('id');
 
-            console.log('deleteConfirmed', deleteConfirm?.length);
+                console.log('deleteConfirmed', deleteConfirm?.length);
+            }
         }
     }
     console.log('updated', updated.length);
